% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ire.R
\docType{data}
\name{ire}
\alias{ire}
\title{Ireland's map}
\format{
A data frame with 34 rows and 84 variables:
}
\usage{
ire
}
\description{
A dataset containing geometry to plot Ireland's map
}
\keyword{datasets}
