% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merged_rk_data.R
\docType{data}
\name{merged_rk_data}
\alias{merged_rk_data}
\title{Ireland's covid19 information}
\format{
A data frame with 245196 rows and 8 variables:
}
\usage{
merged_rk_data
}
\description{
A dataset containing Ireland's monthly covid notices
}
\keyword{datasets}
