% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rk_grouped.R
\docType{data}
\name{rk_grouped}
\alias{rk_grouped}
\title{Ireland's map}
\format{
A data frame with 245196 rows and 8 variables:
}
\usage{
rk_grouped
}
\description{
A dataset containing Ireland's grouped map
}
\keyword{datasets}
